+++ cypress/support/index.js
@@ -1,3 +1,2 @@
 import "./commands";
 import "../../lib/countly";
-import "cypress-localstorage-commands";